Record practice history in localStorage
And show a basic view of it at /stats
